+++ modules/audit_console_login.sh
@@ -68,7 +68,9 @@ audit_console_login () {
             verbose_message "Consoles to disabled on \"$console_list\"" "set"
             sed "s/tty[0-9].*//g" < "$check_file" | grep '[a-z]' > "$temp_file"
             cat "$temp_file" > "$check_file"
-            rm "$temp_file"
+            if [ -f "$temp_file" ]; then
+              rm "$temp_file"
+            fi
           fi
         else
           if [ "$audit_mode" = 1 ]; then
