+++ modules/audit_logrotate.sh
@@ -39,7 +39,9 @@ audit_logrotate () {
             echo "Removing:  Configuring logrotate"
             sed 's,.*{,$search_string {,' "$check_file" > "$temp_file"
             cat "$temp_file" > "$check_file"
-            rm "$temp_file"
+            if [ -f "$temp_file" ]; then
+              rm "$temp_file"
+            fi
           fi
         else
           if [ "$audit_mode" = 1 ]; then
